Created function for creating term tables using the SKOSMOS API. Added API directory for all code regarding API interactions.
